Basic entity class for Imeon inverter integration (#145778)
