+++ tests/components/imeon_inverter/test_sensor.py
@@ -1,6 +1,6 @@
 """Test the Imeon Inverter sensors."""
 
-from unittest.mock import MagicMock, patch
+from unittest.mock import patch
 
 from syrupy.assertion import SnapshotAssertion
 
@@ -15,15 +15,12 @@ from tests.common import MockConfigEntry, snapshot_platform
 
 async def test_sensors(
     hass: HomeAssistant,
-    mock_imeon_inverter: MagicMock,
     entity_registry: er.EntityRegistry,
     snapshot: SnapshotAssertion,
     mock_config_entry: MockConfigEntry,
 ) -> None:
     """Test the Imeon Inverter sensors."""
-    with patch(
-        "homeassistant.components.imeon_inverter.const.PLATFORMS", [Platform.SENSOR]
-    ):
+    with patch("homeassistant.components.imeon_inverter.PLATFORMS", [Platform.SENSOR]):
         await setup_integration(hass, mock_config_entry)
 
     await snapshot_platform(hass, entity_registry, snapshot, mock_config_entry.entry_id)
